migrate ConsumerReportAcknowledgment
